move EaT biasing functions to their own module
